requirements.txt
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
db-dtypes==1.2.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
numpy==1.26.4
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
protobuf==4.25.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sshtunnel==0.4.0
tzdata==2024.1
urllib3==2.2.1
